+++ src/components/general/conditional-wrapper/conditional-wrapper.vue
@@ -1,0 +1,28 @@
+<template>
+	<component :is="tag" v-if="wrap">
+		<slot />
+	</component>
+	<template v-else>
+		<slot />
+	</template>
+</template>
+
+<script setup>
+defineProps({
+	/**
+	 * Whether to wrap the contents.
+	 */
+	wrap: {
+		type: Boolean,
+		default: true,
+	},
+
+	/**
+	 * The tag to wrap with, if wrapping is active.
+	 */
+	tag: {
+		type: String,
+		default: "div",
+	},
+});
+</script>